requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
crispy-bootstrap4==2024.1
Django==4.2.13
django-allauth==0.63.3
django-crispy-forms==2.2
pillow==10.3.0
sqlparse==0.5.0
typing-extensions==4.12.2